% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csfd-package.R
\docType{package}
\name{csfd-package}
\alias{csfd}
\alias{csfd-package}
\title{csfd: Scrape Data from the Czech and Slovak Film Database (CSFD)}
\description{
Download and scrape CSFD pages. This package is not affiliated, associated, authorized, endorsed by or in any way officially connected to CSFD. Please be polite and use it responsibly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jchrom/csfd}
  \item Report bugs at \url{https://github.com/jchrom/csfd/issues}
}

}
\author{
\strong{Maintainer}: Jakub Chromec \email{chromec.jakub@gmail.com}

}
\keyword{internal}
