% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{csfd_fetch}
\alias{csfd_fetch}
\title{Perform A Scraper Request}
\usage{
csfd_fetch(url, quiet = FALSE)
}
\arguments{
\item{url}{A CSFD URL, even a partial one, e.g. \code{"/film/9834-star-trek-film"}}

\item{quiet}{If \code{TRUE}, request status will not be printed on the console.}
}
\value{
An object of class \link{csfd_scraper}.
}
\description{
Download a CSFD page and return a scraper object.
}
\examples{

\dontrun{

# Partial URL works.
trek <- csfd_fetch("/film/9834-star-trek-film/prehled/")
trek

# Even shorter:
trek <- csfd_fetch("film/9834")
trek

# Extract data like you would extract elements from a list.
trek$summary
trek$reviews
trek$ratings
}
}
