% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{csfd_scraper}
\alias{csfd_scraper}
\title{Initialize A New Scraper}
\usage{
csfd_scraper(resp)
}
\arguments{
\item{resp}{An object of class \verb{<httr2_response>}.}
}
\value{
A \verb{<csfd_scraper>}.
}
\description{
Create a new \verb{<csfd_scraper>} object. It is best to use \code{\link[=csfd_fetch]{csfd_fetch()}} instead
of initializing directly.
}
\details{
The scraper object includes a set of scraper functions, each of which returns
a data frame. Trigger scraping by indexing into the object as you would into
a list.

The selection of scraper functions will vary, depending on the type of page.
For example, user pages don't have \emph{Tags} but film pages often do,
so the \code{tags} scraper will be available if you downloaded a film page but not
if you downloaded a user page.

A scraper may sometimes return a zero-row data frame. For example,
while appropriate for films, \emph{Tags} section may still be missing even from
a film page, if no tag has been added yet to that particular film.
}
\examples{

\dontrun{

trek <- csfd_fetch("/film/9834-star-trek-film/prehled/")
trek

# Extract data
trek$summary
trek$ratings
}
}
