% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{csfd_search}
\alias{csfd_search}
\alias{csfd_search_titles}
\alias{csfd_search_creators}
\alias{csfd_search_ranks}
\alias{csfd_form_browse}
\title{Advanced Search}
\usage{
csfd_search_titles(
  type = NULL,
  conditions = NULL,
  genre = NULL,
  origin = NULL,
  rating = NULL,
  released = NULL,
  fields = NULL,
  page = 1,
  sort = c("rated", "rating_average", "rating_average_asc", "fanclub_count", "year",
    "year_asc"),
  quiet = FALSE
)

csfd_search_creators(
  type = NULL,
  conditions = NULL,
  born = NULL,
  born_in = NULL,
  died = NULL,
  died_in = NULL,
  gender = NULL,
  page = 1,
  sort = c("fanclub_count", "birth_date_asc", "birth_date"),
  quiet = FALSE
)

csfd_search_ranks(
  type = NULL,
  origin = NULL,
  released = NULL,
  fields = NULL,
  genre = NULL,
  page = 1,
  quiet = FALSE
)

csfd_form_browse(form = c("titles", "creators", "ranks"))
}
\arguments{
\item{type, conditions}{A vector of checkbox names/dropdown options.}

\item{genre, origin}{For \code{csfd_search_ranks()}, a string, e.g. \code{"Polsko"}
or \code{"tv_show"}.

For \code{csfd_search_titles()}, a named list. Examples:
\itemize{
\item use \code{list(any = c("Fantasy", "Film-Noir"))} to get Fantasy or Film-Noir
\item use \code{list(all = c("Fantasy", "Film-Noir"))} to get films which are both
Fantasy AND Film-Noir
\item use \code{list(exact = c("Fantasy", "Film-Noir"))} for films which are exactly
Fantasy AND Film-Noir and nothing else
\item use \code{list(not = c("Fantasy", "Film-Noir"))} to exclude Fantasy and Film-Noir
from the selection. Useful in combination with \code{any/all}.
}}

\item{rating}{A vector of two numbers.
\itemize{
\item use \code{c(50, 80)} to get films rated between 50 and 80
\item for \code{rating}, numbers between 0 and 1 also work, e.g. \code{c(0.75, 0.9)} is
equivalent to \code{c(75, 90)}
}}

\item{released}{A vector of 1 to 2 numbers. Examples:
\itemize{
\item use \code{2020} to get films released in 2020 or later
\item use \code{c(1940, 1960)} to get creators born between 1940 and 1960
}}

\item{fields}{A named list of numerical identifiers. Examples:
\itemize{
\item use \code{list(actor = c(483, 486), tag = 800)} to search for titles starring
Leonard Nimoy (\code{tvurce/453}) and William Shatner (\code{tvurce/486}), tagged
"Star Trek" (tag id \code{800}).
}}

\item{page}{Which page to retrieve.}

\item{sort}{Results sorting.}

\item{quiet}{If \code{TRUE}, request status will not be printed on the console.}

\item{born, died}{A vector of date strings. Examples:
\itemize{
\item use \verb{c("01.01.1950", "31.12.2022)} to search creators born between
these dates
\item use \code{"500 BC"} to search creators born 500 BC - useful for authors such
as Euripides
}}

\item{born_in, died_in}{A single country name.}

\item{gender}{One of: \code{"male"}, \code{"female"}, \code{1}, \code{2} or \code{NULL}.}

\item{form}{Which list of form fields to show.}
}
\value{
An object of class \link{csfd_scraper}.
}
\description{
Search pages on CSFD.
\itemize{
\item use \code{csfd_search_titles()} to search for \href{https://www.csfd.cz/podrobne-vyhledavani/}{titles}
\item use \code{csfd_search_creators()} to search for \href{https://www.csfd.cz/podrobne-vyhledavani/tvurci/}{creators}
\item use \code{csfd_search_ranks()} to search \href{https://www.csfd.cz/zebricky/vlastni-vyber/}{rankings}
}

Browse available form fields using \code{\link[=csfd_form_browse]{csfd_form_browse()}}, e.g. \code{csfd_form_browse("titles")}.
}
\section{Results}{


The result of the search call is a scraper object with two items:
\enumerate{
\item \code{results}: A data frame with search results, including unique identifier.
\item \code{paginator}: A data frame with \code{next_url} column, which redirects
to the next page of results; if a missing value, it indicates there are
no more results.
}
}

\examples{

\dontrun{

# Search for films and TV shows made in the US only, either drama or sci-fi
# (and possibly other things except horror), rated at least 80 and made
# between 2000 and 2022 (included).
re <- csfd_search_titles(
  type = c("film", "tv_show"),
  origin = list(exact = "USA"),
  genre = list(any = c("Drama", "Sci-Fi"), not = c("horor")),
  rating = 80,
  released = c(2000, 2022)
)

re$results

# As there are more results than would fit a single page, retrieve the next
# page URL and use csfd_fetch() to get those.
csfd_fetch(re$paginator$next_url)
}
}
