% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\docType{data}
\name{entities}
\alias{entities}
\title{Entities}
\format{
A named list with component options.
}
\usage{
entities
}
\description{
A named list of various interactive components on the site. These may include
dropdown menus, region names, checkbox names etc. Mostly for internal use,
but may be useful when building a query for \code{\link[=csfd_search]{csfd_search()}}.
}
\details{
Populated \code{.onLoad()}.
}
\keyword{datasets}
