% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{csfd_fetch}
\alias{csfd_fetch}
\title{Issue A Scrape Request}
\usage{
csfd_fetch(url, quiet = FALSE)
}
\arguments{
\item{url}{A CSFD URL, incl. a partial one, e.g. \code{"/film/9834-star-trek-film"}}

\item{quiet}{If \code{TRUE}, request status will not be printed on the console.

This indicates whether the request is being executed or has been completed
and with what status. It also prints out the response URL which may be
different from the request URL, because the server always returns a complete
URL rather than just the resource identifier.}
}
\value{
An object of class \link{csfd_scraper}.
}
\description{
Download a page from the \href{https://www.csfd.cz}{Czech and Slovak Film Database}
(CSFD).
}
\details{
This function fetches HTML from a single URL. The resulting object contains
a set of fields for scraping different parts of the page. For all available
scrapers, see \link{csfd_scraper_list}.

Note that rate is throttled at 15 requests per minute.
}
\examples{

\dontrun{

# Partial URL works:
trek79 <- csfd_fetch("/film/9834-star-trek-film/prehled/")
trek79

# Even shorter:
trek79 <- csfd_fetch("film/9834")
trek79

# Extract data like you would extract elements from a list:
trek79$details
trek79$reviews
trek79$ratings

# Careful with TV show seasons, they have two different identifiers - you
# want the second one. For example:
# /film/68990-star-trek-hluboky-vesmir-devet/494010-serie-1/prehled/
ds9 <- csfd_fetch("film/494010")
}
}
