% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{csfd_scraper}
\alias{csfd_scraper}
\title{Initialize A New Scraper}
\usage{
csfd_scraper(resp)
}
\arguments{
\item{resp}{An object of class \link[httr2:response]{httr2::response}.}
}
\value{
A \verb{<csfd_scraper>}.
}
\description{
Create a new \verb{<csfd_scraper>} object. This is mostly for debugging purposes,
normally it is best to use \code{\link[=csfd_fetch]{csfd_fetch()}}.
}
\details{
The scraper object includes a set of scraper functions, each of which returns
a data frame. Trigger scraping by indexing into the object as you would into
a list.

The selection of scraper functions will vary, depending on what page you
have fetched. For example, user pages don't have tags but film pages often
do. As a result, \code{tags} scraper will be available if you downloaded a \verb{/film}
page, but not if you downloaded an \verb{/uzivatel} page.

A scraper may return a data frame with no rows. For example, while appropriate
for films, tags may still be missing even from a film page, if no tag has
been added yet to that particular film.
}
