% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper-list.R
\docType{data}
\name{csfd_scraper_list}
\alias{csfd_scraper_list}
\title{CSFD Scraper List}
\format{
A named list with scraper functions.
}
\usage{
csfd_scraper_list
}
\description{
All available scrapers, indexed by name and pattern.
}
\keyword{datasets}
