% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{csfd_search}
\alias{csfd_search}
\title{Advanced Search}
\usage{
csfd_search(
  checkbox = NULL,
  genre = NULL,
  origin = NULL,
  rating = NULL,
  year = NULL,
  fields = NULL,
  page = 1,
  sort = c("rated", "rating_average", "rating_average_asc", "fanclub_count", "year",
    "year_asc"),
  quiet = FALSE
)
}
\arguments{
\item{checkbox}{A vector of checkbox names to check.}

\item{genre, origin}{A named list of choices.
\itemize{
\item use \code{list(any = c("Fantasy", "Film-Noir"))} to get Fantasy or Film-Noir
\item use \code{list(all = c("Fantasy", "Film-Noir"))} to get films which are both
Fantasy AND Film-Noir
\item use \code{list(exact = c("Fantasy", "Film-Noir"))} for films which are exactly
Fantasy AND Film-Noir and nothing else
\item use \code{list(not = c("Fantasy", "Film-Noir"))} to exclude Fantasy and Film-Noir
from the selection. Useful in combination with \code{any/all}.
}

Same logic applies to origin.}

\item{rating, year}{A list of 2 numbers.
\itemize{
\item use \code{c(50, 80)} to get films rated between 50 and 80
\item use \code{2020} to get films released in 2020 or later
\item for \code{rating}, numbers between 0 and 1 also work, e.g. \code{c(0.75, 0.9)} is
equivalent to \code{c(75, 90)}
}}

\item{fields}{A named list of entity identifiers. See \strong{Details.}}

\item{page}{Which page to retrieve.}

\item{sort}{Results sorting.}

\item{quiet}{If \code{TRUE}, request status will not be printed on the console.}
}
\value{
An object of class \link{csfd_scraper}.
}
\description{
Execute \href{https://www.csfd.cz/podrobne-vyhledavani}{advanced search} with your
own parameters.
}
\details{
Searching for creators and tags is a bit cumbersome, because they require
special formatting. To search for a creator, use \verb{id:First Last (nar. YYYY, Country)}, where:
\itemize{
\item \code{id} is a numerical identifier (e.g. \code{453} in \code{tvurce/453-leonard-nimoy})
\item \verb{First Last} should be a full name, e.g. \code{"Robert Downey Jr."}
\item \code{YYYY} is a year of birth
\item \code{Country} is a country of origin, e.g. \code{"USA"}
}

To search a tag, use \code{id:Name}, where:
\itemize{
\item \code{id} is a numerical identifier
\item \code{name} is a full name of the tag, e.g. \code{"Star Trek"}
}

This information is included on creator and film pages, and can be scraped
using \code{\link[=csfd_fetch]{csfd_fetch()}}.

Examples:\preformatted{fields = list(
  actor = c(
    "453:Leonard Nimoy (nar. 1931, USA)",
    "486:William Shatner (nar. 1931, Kanada)"
  ),
  tag = c("17:budoucnost", "800:Star Trek")
)

csfd_search(checkbox = "film", fields = fields)
}

To see all available options, see \link{entities}. For example, to get a list
of available checkboxes, call \code{entities$checkbox}.
}
\examples{

\dontrun{

# Search for films and TV shows made in the US only, either drama or sci-fi,
# rated at least 80 and made between 2000 and 2022 (included).
search <- csfd_search(
  checkbox = c("film", "tv_show"),
  origin = list(exact = "USA"),
  genre = list(any = c("Drama", "Sci-Fi")),
  rating = 80,
  year = c(2000, 2022)
)

search$results
}
}
