% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{csfd_search}
\alias{csfd_search}
\alias{csfd_search_titles}
\alias{csfd_search_creators}
\alias{csfd_search_ranks}
\alias{csfd_search_options}
\title{Advanced Search}
\usage{
csfd_search_titles(
  type = NULL,
  additional = NULL,
  genre = NULL,
  origin = NULL,
  released = NULL,
  rating = NULL,
  fields = NULL,
  page = 1,
  sort = c("rated", "rating_average", "rating_average_asc", "fanclub_count", "year",
    "year_asc"),
  quiet = FALSE
)

csfd_search_creators(
  type = NULL,
  additional = NULL,
  born = NULL,
  born_in = NULL,
  died = NULL,
  died_in = NULL,
  gender = NULL,
  page = 1,
  sort = c("fanclub_count", "birth_date_asc", "birth_date"),
  quiet = FALSE
)

csfd_search_ranks(
  type = NULL,
  origin = NULL,
  released = NULL,
  actor = NULL,
  director = NULL,
  genre = NULL,
  page = 1,
  quiet = FALSE
)

csfd_search_options(domain, element, choices = Inf, single = FALSE)
}
\arguments{
\item{type, additional}{A vector of checkbox names/dropdown options.}

\item{genre, origin}{For \code{csfd_search_ranks()}, a string, e.g. \code{"Polsko"}
or \code{"tv_show"}.

For \code{csfd_search_titles()}, a named list. Examples:
\itemize{
\item use \code{list(any = c("Fantasy", "Film-Noir"))} to get Fantasy or Film-Noir
\item use \code{list(all = c("Fantasy", "Film-Noir"))} to get films which are both
Fantasy AND Film-Noir
\item use \code{list(exact = c("Fantasy", "Film-Noir"))} for films which are exactly
Fantasy AND Film-Noir and nothing else
}

If you use either \code{any} or \code{all}, you can combine it with \code{not} like so:
\itemize{
\item use \code{list(not = c("Horor", "Sci-fi"))} to exclude Horror and Sci-fi
genres from the selection.
}}

\item{released, rating}{A vector of 1 to 2 integers (or \code{NULL}). Examples:
\itemize{
\item use \code{2020} to get films released in 2020 or later
\item use \code{c(50, 80)} to get films rated between 50\% and 80\%
}}

\item{fields}{A named list of numerical identifiers. Examples:
\itemize{
\item use \code{list(actor = c(483, 486), tag = 800)} to search for titles starring
Leonard Nimoy (\code{tvurce/453}) and William Shatner (\code{tvurce/486}), tagged
"Star Trek" (tag id \code{800}).
}}

\item{page}{Which page to retrieve.}

\item{sort}{How to sort results.}

\item{quiet}{If \code{TRUE}, request status will not be printed on the console.}

\item{born, died}{A vector of date strings. Examples:
\itemize{
\item use \verb{c("01.01.1950", "31.12.2022)} to search creators born between
these dates
\item use \code{"500 BC"} to search creators born 500 BC - useful for creators such
as Euripides
}}

\item{born_in, died_in}{A single country name.}

\item{gender}{One of: \code{"male"}, \code{"female"}, \code{1}, \code{2} or \code{NULL}.}

\item{actor, director}{A vector of integer identifiers, obtained from creator's
page URL.}

\item{domain, element, choices}{Useful for displaying options for search forms.
Example:
\itemize{
\item use \code{csfd_search_options("title", "genre", Inf)} for a complete list
of valid genres
\item use \code{csfd_search_options("title", "origin", "USA")} to see if \code{"USA"}
is a valid origin
}}

\item{single}{If \code{TRUE}, only vectors of length 1 will be accepted.}
}
\value{
An object of class \link{csfd_scraper} or, in case of \code{csfd_search_options()},
a list.
}
\description{
Search pages on CSFD.
\itemize{
\item use \code{csfd_search_titles()} to search for \href{https://www.csfd.cz/podrobne-vyhledavani/}{titles}
\item use \code{csfd_search_creators()} to search for \href{https://www.csfd.cz/podrobne-vyhledavani/tvurci/}{creators}
\item use \code{csfd_search_ranks()} to search \href{https://www.csfd.cz/zebricky/vlastni-vyber/}{rankings}
\item use \code{csfd_search_options()} to review search parameters
}
}
\section{Results}{


The result of the search call is a \link{csfd_scraper} object with two items:
\enumerate{
\item \code{results}: A data frame with search results, including unique identifier.
\item \code{paginator}: A data frame with \code{next_url} column, which redirects
to the next page of results; if a missing value, it indicates there are
no more results.
}
}

\examples{

\dontrun{

# Search for films and TV shows made in the US only, either drama or sci-fi
# (and possibly other things except horror), rated at least 80 and made
# between 2000 and 2022 (included).
re <- csfd_search_titles(
  type = c("film", "tv_show"),
  origin = list(exact = "USA"),
  genre = list(any = c("Drama", "Sci-Fi"), not = c("horor")),
  rating = 80,
  released = c(2000, 2022)
)

re$results

# As there are more results than would fit a single page, retrieve the next
# page URL and use csfd_fetch() to get those.
re$paginator$next_url \%>\% csfd_fetch()

# Search rankings of movies filmed in Asia between 1980-2022, starring Donnie
# Yen and Michelle Yeoh.
re <- csfd_search_ranks(
  type = "film",
  origin = "Asie",
  released = c(1980, 2022),
  # This vector does not have to be named, but it's nice for clarity.
  actor = c("Donnie Yen" = 39, "Michelle Yeoh" = 1818),
  genre = "Akční"
)

re$results
}
}
